refactor: split cli/ceo.py into 6 modules, reduce CC, fix import cycle
Split the 1902-line ceo.py god file into focused modules:
- _path_resolver.py: project path resolution and materialization
- _mode_handlers.py: mode detection, review/qa dispatchers
- _task_builder.py: CEO task string construction
- _ceo_dispatch.py: CEO tailer start/stop
- run.py: cmd_run, heartbeat loop, chain modes

Break the insights→registry import cycle by moving
discover_projects() to registry.py with a deprecated wrapper
in insights.py.

Update test mock patch targets to match new module locations
(from X import Y creates a local binding — patches must target
the consuming module, not the defining module).

Closes #917
